refactor: flatten modules and cleanup
