+++ verb_graph.py
